requirements.txt
streamlit
langchain
together
python-dotenv
whisper
gradio
langchain-community
pypdf
sentence-transformers
faiss-cpu
youtube-transcript-api